reupdate

+++ back-end/controllers/adminC.py
@@ -9,8 +9,13 @@ from models.accountDB import AccountType
 from flask_jwt_extended import create_access_token, jwt_required, get_jwt_identity, get_jwt
 
 def getUsers():
-    accounts = Account.query.all()
+    accounts = Account.query.with_entities(Account.account_id, 
+                                           Account.email, 
+                                           Account.password, 
+                                           Account.role).all()
+    
     accounts_json = json.dumps(accounts, cls=AlchemyEncoder)
+
     return jsonify({"accounts": accounts_json})
 
 class AddAccount(Resource):
@@ -31,7 +36,7 @@ class AddAccount(Resource):
             return {'msg': 'Account not found'}, 400
         
         if account.role != AccountType.admin:
-            return {'msg': 'Access forbidden: Only admins can add flights'}, 400
+            return {'msg': 'Access forbidden: Only admins can add accounts'}, 400
 
         data  = AddAccount.parser.parse_args()
         email = data['email']
@@ -52,15 +57,60 @@ class DeleteAccount(Resource):
     parser = reqparse.RequestParser()
     parser.add_argument('email', type=str, required=True, help="This field cannot be left blank")
 
+    @jwt_required()
     def get():
         return getUsers()
     
+    @jwt_required()
     def post():
+
+        current_user_id = get_jwt_identity()
+        account = Account.find_account_id(current_user_id)
+
+        if not account:
+            return {'msg': 'Account not found'}, 400
+        
+        if account.role != AccountType.admin:
+            return {'msg': 'Access forbidden: Only admins can delete accounts'}, 400
+        
         data = DeleteAccount.parser.parse_args()
         email = data['email']
         user = Account.find_email(email)
         user.delete_from_db()
+
+class EditAccount(Resource):
+    parser = reqparse.RequestParser()
+    parser.add_argument('account_id', type=int, required=True, help="This field cannot be left blank")
+    parser.add_argument('email', type=str, required=True, help="This field cannot be left blank")
+    parser.add_argument('password', type=str, required=True, help="This field cannot be left blank")
+
+    @jwt_required()
+    def get(self):
+        return getUsers()
+    
+    @jwt_required()
+    def post(self):
+        current_user_id = get_jwt_identity()
+        account = Account.find_account_id(current_user_id)
+
+        if not account:
+            return {'msg': 'Account not found'}, 400
         
+        if account.role != AccountType.admin:
+            return {'msg': 'Access forbidden: Only admins can add accounts'}, 400
+        
+
+        data = EditAccount.parser.parse_args()
+        account_id = data['account_id']
+        email = data['email']
+        password = data['password']
+
+        user = Account.find_account_id(account_id)
+        user.email = email
+        user.password = generate_password_hash(password)
+        db.session.commit()
+        return getUsers()
+
